Rename promise_executor => executor
