+++ src/webcore/mod.rs
@@ -22,7 +22,7 @@ pub mod promise;
 pub mod promise_future;
 
 #[cfg(feature = "futures")]
-pub mod promise_executor;
+pub mod executor;
 
 #[cfg(feature = "nightly")]
 pub mod void {
